🧹 move asset eol to core
