+++ providers/core/resources/asset_eol.go
@@ -1,54 +1,51 @@
+// Copyright (c) Mondoo, Inc.
+// SPDX-License-Identifier: BUSL-1.1
+
 package resources
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"github.com/rs/zerolog/log"
 	"go.mondoo.com/cnquery/v9/llx"
-	"go.mondoo.com/cnquery/v9/providers-sdk/v1/inventory"
 	"go.mondoo.com/cnquery/v9/providers-sdk/v1/plugin"
 	"go.mondoo.com/cnquery/v9/providers-sdk/v1/resources"
 	"go.mondoo.com/cnquery/v9/providers-sdk/v1/upstream/mvd"
-	"go.mondoo.com/cnquery/v9/providers/vsphere/connection"
+	"go.mondoo.com/cnquery/v9/utils/multierr"
 	"time"
 )
 
-// convertPlatform2VulnPlatform converts the motor platform.Platform to the
-// platform object we use for vulnerability data
-// TODO: we need to harmonize the platform objects
-func convertPlatform2VulnPlatform(pf *inventory.Platform) *mvd.Platform {
-	if pf == nil {
-		return nil
-	}
-	return &mvd.Platform{
-		Name:    pf.Name,
-		Release: pf.Version,
-		Build:   pf.Build,
-		Arch:    pf.Arch,
-		Title:   pf.Title,
-		Labels:  pf.Labels,
-	}
-}
-
 func initAssetEol(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error) {
-	conn := runtime.Connection.(*connection.VsphereConnection)
-	platform := conn.Asset().Platform
-	eolPlatform := convertPlatform2VulnPlatform(platform)
+	pkgs, err := CreateResource(runtime, "asset", nil)
+	if err != nil {
+		return nil, nil, multierr.Wrap(err, "cannot get asset resource")
+	}
+	asset := pkgs.(*mqlAsset)
+
+	labels := map[string]string{}
+	for k, v := range asset.Labels.Data {
+		labels[k] = v.(string)
+	}
+
+	eolPlatform := &mvd.Platform{
+		Name:    asset.Platform.Data,
+		Release: asset.Version.Data,
+		Build:   asset.Build.Data,
+		Arch:    asset.Arch.Data,
+		Title:   asset.Title.Data,
+		Labels:  labels,
+	}
 
 	mcc := runtime.Upstream
 	if mcc == nil || mcc.ApiEndpoint == "" {
 		return nil, nil, resources.MissingUpstreamError{}
 	}
 
-	scannerClient, err := newAdvisoryScannerHttpClient(mcc.ApiEndpoint, mcc.Plugins, mcc.HttpClient)
+	// get new mvd client
+	scannerClient, err := mvd.NewAdvisoryScannerClient(mcc.ApiEndpoint, mcc.HttpClient, mcc.Plugins...)
 	if err != nil {
 		return nil, nil, err
 	}
-
-	data, _ := json.Marshal(eolPlatform)
-	fmt.Println(string(data))
 
 	eolInfo, err := scannerClient.IsEol(context.Background(), eolPlatform)
 	if err != nil {
@@ -71,12 +68,9 @@ func initAssetEol(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[st
 		eolDate = &llx.NeverFutureTime
 	}
 
-	res := mqlAssetEol{
-		MqlRuntime: runtime,
-		DocsUrl:    plugin.TValue[string]{Data: eolInfo.DocsUrl, State: plugin.StateIsSet},
-		ProductUrl: plugin.TValue[string]{Data: eolInfo.ProductUrl, State: plugin.StateIsSet},
-		Date:       plugin.TValue[*time.Time]{Data: eolDate, State: plugin.StateIsSet},
-	}
+	args["docsUrl"] = llx.StringData(eolInfo.DocsUrl)
+	args["productUrl"] = llx.StringData(eolInfo.ProductUrl)
+	args["date"] = llx.TimeDataPtr(eolDate)
 
-	return nil, &res, nil
+	return args, nil, nil
 }
