+++ providers/core/resources/core.lr
@@ -48,6 +48,16 @@ asset @defaults("name platform version") {
   labels map[string]string
 }
 
+// Information about the assets platform end-of-life.
+asset.eol @defaults("date") {
+  // Documentation URL
+  docsUrl string
+  // Product URL
+  productUrl string
+  // End-of-Life date
+  date time
+}
+
 // Date and time functions
 time {
   // The current time on the local system
