+++ providers/os/resources/os.lr.go
@@ -22,21 +22,17 @@ func init() {
 			// to override args, implement: initAsset(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
 			Create: createAsset,
 		},
-		"asset.eol": {
-			Init: initAssetEol,
-			Create: createAssetEol,
-		},
 		"mondoo.eol": {
 			// to override args, implement: initMondooEol(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
 			Create: createMondooEol,
 		},
-		"platform.eol": {
-			Init: initPlatformEol,
-			Create: createPlatformEol,
-		},
 		"platform": {
 			// to override args, implement: initPlatform(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
 			Create: createPlatform,
+		},
+		"platform.eol": {
+			// to override args, implement: initPlatformEol(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
+			Create: createPlatformEol,
 		},
 		"platform.advisories": {
 			// to override args, implement: initPlatformAdvisories(runtime *plugin.Runtime, args map[string]*llx.RawData) (map[string]*llx.RawData, plugin.Resource, error)
@@ -473,15 +469,6 @@ var getDataFields = map[string]func(r plugin.Resource) *plugin.DataRes{
 	"asset.vulnerabilityReport": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlAsset).GetVulnerabilityReport()).ToDataRes(types.Dict)
 	},
-	"asset.eol.docsUrl": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlAssetEol).GetDocsUrl()).ToDataRes(types.String)
-	},
-	"asset.eol.productUrl": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlAssetEol).GetProductUrl()).ToDataRes(types.String)
-	},
-	"asset.eol.date": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlAssetEol).GetDate()).ToDataRes(types.Time)
-	},
 	"mondoo.eol.product": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlMondooEol).GetProduct()).ToDataRes(types.String)
 	},
@@ -491,6 +478,9 @@ var getDataFields = map[string]func(r plugin.Resource) *plugin.DataRes{
 	"mondoo.eol.date": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlMondooEol).GetDate()).ToDataRes(types.Time)
 	},
+	"platform.vulnerabilityReport": func(r plugin.Resource) *plugin.DataRes {
+		return (r.(*mqlPlatform).GetVulnerabilityReport()).ToDataRes(types.Dict)
+	},
 	"platform.eol.docsUrl": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlPlatformEol).GetDocsUrl()).ToDataRes(types.String)
 	},
@@ -499,9 +489,6 @@ var getDataFields = map[string]func(r plugin.Resource) *plugin.DataRes{
 	},
 	"platform.eol.date": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlPlatformEol).GetDate()).ToDataRes(types.Time)
-	},
-	"platform.vulnerabilityReport": func(r plugin.Resource) *plugin.DataRes {
-		return (r.(*mqlPlatform).GetVulnerabilityReport()).ToDataRes(types.Dict)
 	},
 	"platform.advisories.cvss": func(r plugin.Resource) *plugin.DataRes {
 		return (r.(*mqlPlatformAdvisories).GetCvss()).ToDataRes(types.Resource("audit.cvss"))
@@ -1915,22 +1902,6 @@ var setDataFields = map[string]func(r plugin.Resource, v *llx.RawData) bool {
 		r.(*mqlAsset).VulnerabilityReport, ok = plugin.RawToTValue[interface{}](v.Value, v.Error)
 		return
 	},
-	"asset.eol.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-			r.(*mqlAssetEol).__id, ok = v.Value.(string)
-			return
-		},
-	"asset.eol.docsUrl": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlAssetEol).DocsUrl, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"asset.eol.productUrl": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlAssetEol).ProductUrl, ok = plugin.RawToTValue[string](v.Value, v.Error)
-		return
-	},
-	"asset.eol.date": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlAssetEol).Date, ok = plugin.RawToTValue[*time.Time](v.Value, v.Error)
-		return
-	},
 	"mondoo.eol.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
 			r.(*mqlMondooEol).__id, ok = v.Value.(string)
 			return
@@ -1947,6 +1918,14 @@ var setDataFields = map[string]func(r plugin.Resource, v *llx.RawData) bool {
 		r.(*mqlMondooEol).Date, ok = plugin.RawToTValue[*time.Time](v.Value, v.Error)
 		return
 	},
+	"platform.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
+			r.(*mqlPlatform).__id, ok = v.Value.(string)
+			return
+		},
+	"platform.vulnerabilityReport": func(r plugin.Resource, v *llx.RawData) (ok bool) {
+		r.(*mqlPlatform).VulnerabilityReport, ok = plugin.RawToTValue[interface{}](v.Value, v.Error)
+		return
+	},
 	"platform.eol.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
 			r.(*mqlPlatformEol).__id, ok = v.Value.(string)
 			return
@@ -1961,14 +1940,6 @@ var setDataFields = map[string]func(r plugin.Resource, v *llx.RawData) bool {
 	},
 	"platform.eol.date": func(r plugin.Resource, v *llx.RawData) (ok bool) {
 		r.(*mqlPlatformEol).Date, ok = plugin.RawToTValue[*time.Time](v.Value, v.Error)
-		return
-	},
-	"platform.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-			r.(*mqlPlatform).__id, ok = v.Value.(string)
-			return
-		},
-	"platform.vulnerabilityReport": func(r plugin.Resource, v *llx.RawData) (ok bool) {
-		r.(*mqlPlatform).VulnerabilityReport, ok = plugin.RawToTValue[interface{}](v.Value, v.Error)
 		return
 	},
 	"platform.advisories.__id": func(r plugin.Resource, v *llx.RawData) (ok bool) {
@@ -4261,60 +4232,6 @@ func (c *mqlAsset) GetVulnerabilityReport() *plugin.TValue[interface{}] {
 	})
 }
 
-// mqlAssetEol for the asset.eol resource
-type mqlAssetEol struct {
-	MqlRuntime *plugin.Runtime
-	__id string
-	// optional: if you define mqlAssetEolInternal it will be used here
-	DocsUrl plugin.TValue[string]
-	ProductUrl plugin.TValue[string]
-	Date plugin.TValue[*time.Time]
-}
-
-// createAssetEol creates a new instance of this resource
-func createAssetEol(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlAssetEol{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("asset.eol", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlAssetEol) MqlName() string {
-	return "asset.eol"
-}
-
-func (c *mqlAssetEol) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlAssetEol) GetDocsUrl() *plugin.TValue[string] {
-	return &c.DocsUrl
-}
-
-func (c *mqlAssetEol) GetProductUrl() *plugin.TValue[string] {
-	return &c.ProductUrl
-}
-
-func (c *mqlAssetEol) GetDate() *plugin.TValue[*time.Time] {
-	return &c.Date
-}
-
 // mqlMondooEol for the mondoo.eol resource
 type mqlMondooEol struct {
 	MqlRuntime *plugin.Runtime
@@ -4376,6 +4293,52 @@ func (c *mqlMondooEol) GetDate() *plugin.TValue[*time.Time] {
 	})
 }
 
+// mqlPlatform for the platform resource
+type mqlPlatform struct {
+	MqlRuntime *plugin.Runtime
+	__id string
+	// optional: if you define mqlPlatformInternal it will be used here
+	VulnerabilityReport plugin.TValue[interface{}]
+}
+
+// createPlatform creates a new instance of this resource
+func createPlatform(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
+	res := &mqlPlatform{
+		MqlRuntime: runtime,
+	}
+
+	err := SetAllData(res, args)
+	if err != nil {
+		return res, err
+	}
+
+	// to override __id implement: id() (string, error)
+
+	if runtime.HasRecording {
+		args, err = runtime.ResourceFromRecording("platform", res.__id)
+		if err != nil || args == nil {
+			return res, err
+		}
+		return res, SetAllData(res, args)
+	}
+
+	return res, nil
+}
+
+func (c *mqlPlatform) MqlName() string {
+	return "platform"
+}
+
+func (c *mqlPlatform) MqlID() string {
+	return c.__id
+}
+
+func (c *mqlPlatform) GetVulnerabilityReport() *plugin.TValue[interface{}] {
+	return plugin.GetOrCompute[interface{}](&c.VulnerabilityReport, func() (interface{}, error) {
+		return c.vulnerabilityReport()
+	})
+}
+
 // mqlPlatformEol for the platform.eol resource
 type mqlPlatformEol struct {
 	MqlRuntime *plugin.Runtime
@@ -4428,52 +4391,6 @@ func (c *mqlPlatformEol) GetProductUrl() *plugin.TValue[string] {
 
 func (c *mqlPlatformEol) GetDate() *plugin.TValue[*time.Time] {
 	return &c.Date
-}
-
-// mqlPlatform for the platform resource
-type mqlPlatform struct {
-	MqlRuntime *plugin.Runtime
-	__id string
-	// optional: if you define mqlPlatformInternal it will be used here
-	VulnerabilityReport plugin.TValue[interface{}]
-}
-
-// createPlatform creates a new instance of this resource
-func createPlatform(runtime *plugin.Runtime, args map[string]*llx.RawData) (plugin.Resource, error) {
-	res := &mqlPlatform{
-		MqlRuntime: runtime,
-	}
-
-	err := SetAllData(res, args)
-	if err != nil {
-		return res, err
-	}
-
-	// to override __id implement: id() (string, error)
-
-	if runtime.HasRecording {
-		args, err = runtime.ResourceFromRecording("platform", res.__id)
-		if err != nil || args == nil {
-			return res, err
-		}
-		return res, SetAllData(res, args)
-	}
-
-	return res, nil
-}
-
-func (c *mqlPlatform) MqlName() string {
-	return "platform"
-}
-
-func (c *mqlPlatform) MqlID() string {
-	return c.__id
-}
-
-func (c *mqlPlatform) GetVulnerabilityReport() *plugin.TValue[interface{}] {
-	return plugin.GetOrCompute[interface{}](&c.VulnerabilityReport, func() (interface{}, error) {
-		return c.vulnerabilityReport()
-	})
 }
 
 // mqlPlatformAdvisories for the platform.advisories resource
